requirements.txt
uvicorn
fastapi[all] 
fastapi-mail
jinja2
pydantic
python-dotenv